fix: [lw-10691]: fix sonarcloud issues
