+++ packages/staking/src/features/overview/OverviewPopup.tsx
@@ -3,7 +3,6 @@ import { Box, Flex, Text } from '@input-output-hk/lace-ui-toolkit';
 import { Wallet } from '@lace/cardano';
 import { useObservable } from '@lace/common';
 import { Skeleton } from 'antd';
-import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DelegationCard } from '../DelegationCard';
 import { useOutsideHandles } from '../outside-handles-provider';
@@ -29,6 +28,7 @@ export const OverviewPopup = () => {
     isSharedWallet,
     openExternalLink,
     setIsRegisterAsDRepModalVisible,
+    useRewardAccountsData,
   } = useOutsideHandles();
   const rewardAccounts = useObservable(inMemoryWallet.delegation.rewardAccounts$);
   const protocolParameters = useObservable(inMemoryWallet.protocolParameters$);
@@ -40,25 +40,8 @@ export const OverviewPopup = () => {
 
   const totalCoinBalance = balancesBalance?.total?.coinBalance || '0';
 
-  const rewardAccountsWithRegisteredStakeCreds = rewardAccounts?.filter(
-    ({ credentialStatus }) => Wallet.Cardano.StakeCredentialStatus.Registered === credentialStatus,
-  );
-  const showRegisterAsDRepBanner =
-    rewardAccountsWithRegisteredStakeCreds?.length > 0 &&
-    !rewardAccountsWithRegisteredStakeCreds.some(({ dRepDelegatee }) => dRepDelegatee);
-
-  const poolIdToRewardAccountMap = useMemo(
-    () =>
-      new Map(
-        rewardAccountsWithRegisteredStakeCreds?.map((rewardAccount) => {
-          const { delegatee } = rewardAccount;
-          const delagationInfo = delegatee?.nextNextEpoch || delegatee?.nextEpoch || delegatee?.currentEpoch;
-
-          return [delagationInfo?.id.toString(), rewardAccount];
-        }),
-      ),
-    [rewardAccountsWithRegisteredStakeCreds],
-  );
+  const { areAllRegisteredStakeKeysWithoutVotingDelegation: showRegisterAsDRepBanner, poolIdToRewardAccountMap } =
+    useRewardAccountsData();
 
   if (
     !totalCoinBalance ||
